fix(amount): 修复金额计算异常并提升精度
问题：
- 传入金额为负数时未做校验，导致计算结果异常
- 金额使用 double 计算存在精度丢失问题

方案：
- 添加金额负数校验，非法输入直接拦截
- 将金额相关字段及计算逻辑由 double 替换为 BigDecimal，避免精度损失

+++ atm/src/main/java/com/example/atm/mapper/AccountMapper.java
@@ -2,6 +2,9 @@ package com.example.atm.mapper;
 
 import com.example.atm.entity.Account;
 import org.apache.ibatis.annotations.*;
+
+import java.math.BigDecimal;
+
 @Mapper
 public interface AccountMapper {
 
@@ -40,7 +43,7 @@ public interface AccountMapper {
         WHERE card = #{card}
     """)
     void updateBalance(@Param("card") String card,
-                       @Param("balance") double balance);
+                       @Param("balance") BigDecimal balance);
 
     @Update("""
         UPDATE account
